Updated config for wardenkms
Lint fixes

Merge conflicts

Added ClientsPool abstraction

Build fixes

Lint fix

Lint fix

Update keychain-sdk/tx_client_pool.go

Code review fixes

Merge branch 'feature/683-keychain-sdk-configure-multiple-nodes-endpoints' of github.com:warden-protocol/wardenprotocol into feature/683-keychain-sdk-configure-multiple-nodes-endpoints

Added removal of processed requests

Lint fix

+++ keychain-sdk/config.go
@@ -55,7 +55,7 @@ type Config struct {
 	BasicConfig
 
 	// ConsensusNodeThreshold represents the number of nodes required to execute a pending key/sign request.
-	ConsensusNodeThreshold uint
+	ConsensusNodeThreshold uint8
 
 	// GRPCURLs is the list of URLs of the gRPC server to connect to.
 	// e.g. "localhost:9090"
